Add support to docker-based execution.
With these changes, if a docker image exists locally with name edge2ai-workshop:latest,
all the commands will run on the docker immage rather than locally. The switch to docker is
transparent to the user and no command syntaxes are changed.

The local base directory is mounted on the docker container so that all the Terraform state
can be stored locally and re-used, interchangeably, by local and docker executions of the scripts.

Local executions can be forced by setting:
export NO_DOCKER=1