Made the FreeIPA UI accessible remotely.
The original implementatation didn't allocate a fixed name and public IP
address for the IPA server, which made it harder to access the UI from
our laptops, requiring fiddling with entries in /etc/hosts.

Now the FreeIPA server has a permanent name of the form ipa.x.w.y.x.nip.com
and an elastic IP so that the name survives stop and restarts.