Improve deployment reliability.
Prior to this change the Terraform scripts ran in the foreground for
as long as it took to complete the deployment. Any hiccups in the network
connection would cause the deployment to fail and be in an incosistent
state.

This change modifies the TF script so that the deployment is launched
with nohup in the background on the VMs. The TF execution finishes after
that and the launch script only monitors the execution remotely. This reduces
the succeptibility of the deployment to network instability.

Furthermore, even if the launch is interrupted, the deployment will continue
to run. Once the launch is executed again, it will resume the monitoring for
the current deployment.