Check for MiNiFi service definition file before trying to modify it.
In newer versions, MiNiFi stopped using service-based daemon and started using systemd configuration files.
This caused the deployment to fail because the launch script was trying to update the old file that
not longer existed.

Added a temporary check for file existence to avoid a hard break.
TODO: This needs improvement going forward to ensure we can add the Python 39 binaries to the MiNiFi
path so that we can use Python processors.