Improvements and fixes to CDSW deployment.
Complete refactoring of the CDSW setup script to make it more robust and easier to maintain.
Changed some of the labs Python methods to support the CDSW setup refactoring.
Changed the call to cdsw_setup.py to use the new command line options.
Added a wait for the IPA server to be ready in the CDP deployment to avoid a racing condition.
Fixed a few issues with the setup clean_all procedure to ensure retries work correctly.

 "This commit does not contain secrets"